requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
schedule==1.2.0
openai==1.3.7
sqlite3
feedparser==6.0.10
newspaper3k==0.2.8
arxiv==1.4.8
scholarly==1.7.11
patents==1.0.0
jieba==0.42.1
wordcloud==1.9.2
matplotlib==3.7.2
pandas==2.1.1
numpy==1.24.3
python-dotenv==1.0.0
gunicorn==21.2.0
APScheduler==3.10.4 